Criação das operações mais básicas, alteração no bit do banco em varias tabelas, criação do middware, rota de login e geração de token

+++ DB/tables/itens_pedidos.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS itens_pedidos (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    pedido_id BIGINT NOT NULL,
-    produto_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    pedido_id INT NOT NULL,
+    produto_id INT NOT NULL,
     quantidade INT NOT NULL CHECK (quantidade > 0),
     preco_unitario DECIMAL(10,2) NOT NULL CHECK (preco_unitario >= 0),
     subtotal DECIMAL(10,2) NOT NULL CHECK (subtotal >= 0),
